+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added scheduling recurring messages using crontab formatted strings
 - added scheduling reminder messages to be sent
 - added listing the scheduled recurring messages
+- added removing a scheduled recurring message
 - can automatically render message links 
 - can automatically assign roles based on reactions to a message
 - can generate Spotify playlists